Added Sequencer.tryNext(n, requiredCapacity) method
This method makes it possible to implement a back-off strategy
in a race free way, if some producers should have some sort of
lower priority and not be allowed to fill up the entire input queue.
This assumes that it's the consumers and not the producers that
are the bottleneck of the system. The method is implemented for
the single producer case also, but does not solve anything there
that could not already be done with the existing API:s.